억제한 것들을 실증하고, 그 과정에서 찾은 동작 변경을 되돌린다
"억제했다" 는 판단이고, 판단은 근거가 남아야 다음 사람이 같은 것을 다시 세우지 않는다. 이번 세션에
내가 끄거나 NOLINT 로 덮은 것들을 하나씩 확인했다.

**ASan report_globals=0 (SmokeTest 전용)** — 네 결함을 일부러 내는 12줄 프로브를 ASan 으로 빌드해
같은 옵션으로 돌렸다. heap-buffer-overflow · heap-use-after-free · stack-buffer-overflow 는 그대로
잡히고 global-buffer-overflow 만 안 잡힌다. 기본 옵션으로는 그 global 도 잡히는 것까지 확인해
프로브가 유효함을 보였다. 잃는 것이 정확히 하나이고, 다른 네 테스트는 기본값이라 그 검사도 그대로다.

**.clang-tidy 에서 끈 검사 셋** — 전수 확인했다. invalid-enum 48건은 열거형 19종이 전부 D3D11/
D3D12/Vulkan SDK 타입이고 sw:: 열거형은 0건. derived-method-shadowing 11건은 전부 swReflectSelf.
std-namespace-modification 10건은 전부 tuple_size/tuple_element/hash/equal_to 를 프로그램 정의 타입에
특수화한 것이라 [namespace.std] 가 허용하는 형태다 — 다만 tuple_size<FormattedValue<T>> 가
integral_constant<uint32,2> 였다. 표준은 size_t 를 요구한다(구조적 바인딩이 우연히 동작했을 뿐).
size_t 로 맞췄다.

**NOLINT 두 곳** — BVHTree3D 의 근거로 든 static_assert( sizeof(float4x4) == 16*sizeof(float32) )
가 MatrixMath.h:471 에 실제로 있다. fixed_string 의 자기대입은 주장만 있고 시험이 없었으므로 두
경로(fs = fs, fs = fs.c_str())를 테스트로 고정했다.

**패널 덤프의 "정수 id 자식 창" 규칙** — 실제 덤프에서 걸리는 창은 Sequencer/00000379 하나뿐이다.
우리 자식 창은 전부 문자열 id 라 가려지지 않는다.

**그리고 이번 세션의 변경 하나가 동작을 바꿨다.** Component 생성자에서 기본값 적용을 걷어내면서
gamedata 의 기반 타입 노드(<SceneComponent> 같은)가 적용될 자리가 사라졌다. 예전 코드도 온전하지는
않았다 — 기반 생성자에서는 가상 getTypeInfo() 가 파생으로 디스패치되지 않아 언제나 Component 노드
하나만 봤고 중간 기반은 한 번도 적용된 적이 없다. 이제 applyTypeDefaults 가 상속 체인을 뿌리 →
파생 순서로 전부 적용한다(파생이 마지막에 덮어쓴다).

EngineTest.ComponentDefaults.BaseTypeDefaultsApplyBeforeDerived 로 고정했고, 고치기 전 코드로
되돌리면 기반 값 세 개가 실패하는 것까지 확인했다. (곁들여: 벡터 기본값의 텍스트 형식은 쉼표
구분이다 — "2,3,4". 공백으로 적으면 파싱이 조용히 실패하고 값이 그대로 남는다.)

검증: Debug·Shipping·ASan 빌드 경고 0, nogpu 5/5 세 구성 전부, lint 6/6,
CoreTest 169 / EngineTest 424 / ReflectionTest 100(+1 스킵) / EditorTest 34,
네 백엔드 × 에디터 [Error] 0건 · 14창/빈 0개, 전부 열기 29창/빈 0개,
ASan 앱 실행(dx12 · vk · 에디터+전체 패널) 보고 0건.

+++ Source/Core/String/formatString.h
@@ -903,12 +903,17 @@ namespace sw
 namespace std
 {
     template <typename T>
-    /** @brief FormattedValue 는 (값, Format) 두 원소입니다. */
-    struct tuple_size<sw::FormattedValue<T>> : integral_constant<uint32, 2>
+    /**
+     * @brief FormattedValue 는 (값, Format) 두 원소입니다.
+     * @details `tuple_size` 특수화는 `integral_constant<size_t, N>` 에서 파생해야 표준을 따른다
+     *          (`std::tuple_size<E>::value` 의 타입이 `size_t` 다). 예전에는 `uint32` 였다 —
+     *          구조적 바인딩이 우연히 동작했을 뿐 [tuple.helper] 의 요구는 아니었다.
+     */
+    struct tuple_size<sw::FormattedValue<T>> : integral_constant<size_t, 2>
     {
     };
 
-    template <uint32 TIndex, typename TType>
+    template <size_t TIndex, typename TType>
     /** @brief 0 은 값 타입, 1 은 Format 입니다. */
     struct tuple_element<TIndex, sw::FormattedValue<TType>>
     {
